feat: modify createdGroup logic

+++ backend/ifit/src/main/java/kr/co/ifit/api/response/AddGroupResponseDTO.java
@@ -7,10 +7,10 @@ import lombok.Setter;
 @Setter
 @Getter
 public class AddGroupResponseDTO {
-    private Integer communityId;
+    private Long communityId;
     private String message;
 
-    public AddGroupResponseDTO(Integer communityId, String message) {
+    public AddGroupResponseDTO(Long communityId, String message) {
         this.communityId = communityId;
         this.message = message;
     }
