+++ backend/ifit/src/main/java/kr/co/ifit/db/repository/GroupRepository.java
@@ -1,6 +1,7 @@
 package kr.co.ifit.db.repository;
 
 import kr.co.ifit.db.entity.Group;
+import kr.co.ifit.db.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,7 +9,7 @@ import java.util.List;
 
 // JPA 자동 구현
 @Repository
-public interface GroupRepository extends JpaRepository<Group, Integer> {
+public interface GroupRepository extends JpaRepository<Group, Long> {
     //  userId로 Group 리스트를 조회하는 메서드 정의
-    List<Group> findByUserId(Integer userId);
+    List<Group> findByUserId(Long userId);
 }
